+++ apps/acnw/next.config.js
@@ -20,7 +20,7 @@ const withMdxFm = require('next-mdx-frontmatter')({
 module.exports = withBundleAnalyzer(
   withMdxFm({
     reactStrictMode: true,
-    transpilePackages: ['ui', 'database', 'amber', '@mui/material'],
+    transpilePackages: ['ui', 'database', 'amber', '@mui/material', '@amber/api'],
     eslint: {
       // Warning: This allows production builds to successfully complete even if
       // your project has ESLint errors.
