+++ apps/acnw/pages/api/send/templates/paymentConfirmation/subject.pug
@@ -1,0 +1,1 @@
+= `ACNW ${year} Payment Confirmation - ${name}`
