+++ apps/acus/pages/api/send/gameConfirmation.ts
@@ -1,10 +1,8 @@
-import { emailer, handleError, JsonError } from '@amber/api'
+import { getEmails, emailer, handleError, JsonError } from '@amber/api'
 import { withApiAuthRequired } from '@auth0/nextjs-auth0'
 import { GameConfirmationBody } from 'amber/utils/apiTypes'
 import { getPlayerPreference } from 'amber/utils/selectValues'
 import { NextApiRequest, NextApiResponse } from 'next'
-
-import { getEmails } from '../_config'
 
 // /api/send/gameConfirmation
 // auth token: required
