+++ packages/api/package.json
@@ -22,11 +22,14 @@
     "graphql-tag": "^2.12.6",
     "isomorphic-fetch": "^3.0.0",
     "json2xls": "^0.1.2",
+    "luxon": "^3.3.0",
     "micro": "^9.4.1",
     "next": "13.4.4",
     "nodemailer": "^6.9.3",
     "pg": "^8.11.0",
-    "stripe": "^12.8.0"
+    "pug": "^3.0.2",
+    "stripe": "^12.8.0",
+    "zod": "^3.21.4"
   },
   "devDependencies": {
     "@graphql-codegen/add": "5.0.0",
@@ -36,6 +39,7 @@
     "@graphql-codegen/typescript-operations": "4.0.0",
     "@types/email-templates": "^10.0.1",
     "@types/isomorphic-fetch": "^0.0.36",
+    "@types/luxon": "^3.3.0",
     "@types/node": "^20.2.5",
     "@types/nodemailer": "^6.4.8",
     "@types/pg": "^8.10.1",
