+++ packages/api/src/index.ts
@@ -3,12 +3,13 @@ export * from './auth'
 export * from './constants'
 export * from './email'
 export * from './getConfig'
+export * from './getConfigHandler'
 export * from './getManagementApiAccessToken'
-export * from './getProfile'
-export * from './graphql'
+export * from './getProfileHandler'
+export * from './graphqlHandler'
 export * from './handleError'
 export * from './query'
-export * from './resetPassword'
+export * from './resetPasswordHandler'
 export * from './stripe'
 export * from './utils'
-export * from './validatePassword'
+export * from './validatePasswordHandler'
