+++ packages/api/src/stripe/getStripeConfig.ts
@@ -7,7 +7,7 @@ import { handleError } from '../handleError'
 // auth token: not required
 // body: {}
 
-export const getStripeConfigRoute = async (req: NextApiRequest, res: NextApiResponse) => {
+export const getStripeConfigHandler = async (req: NextApiRequest, res: NextApiResponse) => {
   const baseUrl = process.env.VERCEL_URL?.startsWith('http')
     ? process.env.VERCEL_URL
     : `https://${process.env.VERCEL_URL}`
