+++ packages/api/src/stripe/types.ts
@@ -1,0 +1,6 @@
+// keep in sync with packages/amber/views/Payment/ElementsForm.tsx
+export type UserPaymentDetails = {
+  userId: number
+  memberId: number | null
+  amount: number
+}
