+++ packages/api/src/utils/index.ts
@@ -1,2 +1,3 @@
+export * from './configuration'
 export * from './stripeHelpers'
 export * from './user'
